+++ internal/app/fiber.go
@@ -2,18 +2,28 @@ package app
 
 import (
 	"fmt"
+	"time"
 
+	"github.com/gofiber/fiber/v2"
 	_ "github.com/maktoobgar/go_template/internal/app/load"
 	g "github.com/maktoobgar/go_template/internal/global"
 	"github.com/maktoobgar/go_template/internal/routers"
-	"github.com/maktoobgar/go_template/pkg/api"
+	"github.com/maktoobgar/go_template/pkg/errors"
 )
 
 var ()
 
 func Fiber() {
 	// Run App
-	app := api.New("Brand New App")
+	app := fiber.New(fiber.Config{
+		Prefork:      true,
+		ErrorHandler: errors.ErrorHandler,
+		ReadTimeout:  time.Duration(time.Second * 5),
+		WriteTimeout: time.Duration(time.Second * 30),
+		IdleTimeout:  time.Duration(time.Minute * 5),
+		AppName:      "Brand New App",
+	})
+
 	g.App = app
 
 	// Router Settings