+++ internal/services/session/session.go
@@ -9,13 +9,9 @@ import (
 	g "github.com/maktoobgar/go_template/internal/global"
 	"github.com/maktoobgar/go_template/internal/models"
 	"github.com/maktoobgar/go_template/pkg/errors"
-	"github.com/maktoobgar/go_template/pkg/errors/messages"
 )
 
 type sessionService struct{}
-
-var errDataNotFound = errors.New(errors.NotFoundStatus, messages.ErrDataNotFound)
-var errDataDuplication = errors.New(errors.NotFoundStatus, messages.ErrDataDuplication)
 
 var instance = &sessionService{}
 
@@ -25,7 +21,7 @@ func (obj *sessionService) get_value(key string) (*models.Session, error) {
 		goqu.Ex{"key": key},
 	).Executor().ScanStruct(sessionObj)
 	if !ok {
-		return nil, errDataNotFound
+		return nil, errors.New(errors.UnauthorizedStatus, errors.Resend, g.Translator.TranslateEN("InvalidSessionID"))
 	}
 
 	return sessionObj, nil
@@ -41,12 +37,12 @@ func (obj *sessionService) Get(key string) ([]byte, error) {
 
 	if session.ExpireDate.Unix() < time.Now().Unix() {
 		obj.Delete(key)
-		return nil, errDataNotFound
+		return nil, errors.New(errors.UnauthorizedStatus, errors.ReSingIn, g.Translator.TranslateEN("ExpiredSessionID"))
 	}
 
 	res, err := hex.DecodeString(session.Value)
 	if err != nil {
-		return nil, errors.New(errors.UnexpectedStatus, err.Error())
+		return nil, errors.New(errors.UnexpectedStatus, errors.ReSingIn, g.Translator.TranslateEN("DecodeFailure"))
 	}
 	return res, nil
 }
@@ -57,7 +53,7 @@ func (obj *sessionService) Get(key string) ([]byte, error) {
 func (obj *sessionService) Set(key string, val []byte, ttl time.Duration) error {
 	v, _ := obj.Get(key)
 	if v != nil {
-		return errDataDuplication
+		return errors.New(errors.InvalidStatus, errors.Resend, g.Translator.TranslateEN("DuplicateSession"))
 	}
 	if key == "" || val == nil {
 		return nil
@@ -74,7 +70,7 @@ func (obj *sessionService) Set(key string, val []byte, ttl time.Duration) error 
 
 	_, err := g.DB.Insert(models.SessionName).Rows(sessions).Executor().Exec()
 	if err != nil {
-		return errors.New(errors.UnexpectedStatus, err.Error())
+		return errors.New(errors.UnexpectedStatus, errors.Report, g.Translator.TranslateEN("CreationSessionFailed"))
 	}
 
 	return nil
@@ -87,7 +83,7 @@ func (obj *sessionService) Delete(key string) error {
 		"key": key,
 	}).Executor().Exec()
 	if err != nil {
-		return err
+		return errors.New(errors.UnexpectedStatus, errors.Report, g.Translator.TranslateEN("DeletionSessionFailed"))
 	}
 
 	return nil
@@ -97,7 +93,7 @@ func (obj *sessionService) Delete(key string) error {
 func (obj *sessionService) Reset() error {
 	_, err := g.DB.Delete(models.SessionName).Executor().Exec()
 	if err != nil {
-		return errors.New(errors.UnexpectedStatus, err.Error())
+		return errors.New(errors.UnexpectedStatus, errors.Report, g.Translator.TranslateEN("ResetSessionFailed"))
 	}
 
 	return nil