+++ pkg/errors/errors.go
@@ -1,14 +1,16 @@
 package errors
 
 import (
+	"fmt"
 	"net/http"
 
-	"github.com/maktoobgar/go_template/pkg/errors/messages"
+	"github.com/gofiber/fiber/v2"
 )
 
 type (
 	serverError struct {
 		code    int
+		action  int
 		message string
 	}
 )
@@ -27,6 +29,16 @@ const (
 	NotAllowedStatus
 )
 
+const (
+	_ int = iota
+	// SignIn in again
+	ReSingIn
+	// Report the problem
+	Report
+	// Correct sent data and request again
+	Resend
+)
+
 var (
 	httpErrors = map[int]int{
 		InvalidStatus:      http.StatusBadRequest,
@@ -41,25 +53,39 @@ func (e serverError) Error() string {
 	return e.message
 }
 
-// Returns httpErrorCode and message of it
-func HttpError(e error) (string, int) {
-	err, ok := e.(serverError)
-	if !ok {
-		return messages.ErrorGeneral, httpErrors[UnexpectedStatus]
+// Returns httpErrorCode, message and action of it
+func HttpError(err error) (code int, action int, message string) {
+	code = http.StatusInternalServerError
+	action = Report
+	message = err.Error()
+
+	if er, ok := err.(serverError); ok {
+		code = httpErrors[er.code]
+		action = er.action
+		message = er.message
 	}
 
-	httpCode, ok := httpErrors[err.code]
-	if !ok {
-		return messages.ErrorGeneralNotFound, httpErrors[InvalidStatus]
-	}
+	return
+}
 
-	return err.message, httpCode
+func ErrorHandler(c *fiber.Ctx, err error) error {
+	// Set Content-Type: application/json; charset=utf-8
+	c.Set(fiber.HeaderContentType, fiber.MIMEApplicationJSONCharsetUTF8)
+
+	code, action, message := HttpError(err)
+
+	return c.Status(code).JSON(map[string]string{
+		"Code":    fmt.Sprint(code),
+		"Action":  fmt.Sprint(action),
+		"Message": message,
+	})
 }
 
 // Creates a new error
-func New(code int, message string) error {
+func New(code int, action int, message string) error {
 	return serverError{
 		code:    code,
+		action:  action,
 		message: message,
 	}
 }